requirements.txt
Django==1.10.1
django-dashing==0.3.1
gunicorn==19.7
pymongo==3.1.1
pprint